refactor: decompose delegate/tools.rs into cohesive submodules
Split the 583-line tools.rs into four focused submodules:
- preflight: hook evaluation, approval resolution, batch grouping
- execution: inline and parallel batch tool runners
- status: lifecycle events, image sentinels, output sanitisation
- recording: outcome recording, context folding, post-flight processing

Additional changes:
- Guard hook arguments to accept only JSON objects (warn on non-object)
- Always run tool output through SafetyLayer (remove JSON bypass)
- Collapse parse_auth_result tests into rstest parameterized test
- Extract make_test_skill/install_skill helpers in skills tests
- Fix poisoned registry test to be non-empty before poisoning
- Replace unwrap() with expect() in auth and skills tests
- Remove unused is_valid_json function and unused imports
